requirements.txt
Celery
Django==2.2.11
dj-database-url==0.5.0
djangorestframework==3.11.0
face_recognition==1.3.0
psycopg2-binary==2.8.3
